refactor: consolidate test cases into ImageProcessTest.cpp
Move SinglePixelResizeTest and StrideMismatchTest from the separate
ImageOpTest.cpp into the existing test/cv/ImageProcessTest.cpp test
suite, and delete ImageOpTest.cpp.

Test suite names updated from cv/image_op/* to cv/image_process/* to
match the existing convention.

Requested by @wangzhaode in review.
